backend/requirements.txt
# Django framework
Django==5.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
django-cors-headers== 4.3.1
gunicorn==22.0.0
djangorestframework-simplejwt==5.3.1
python-decouple
flake8
graphene-django
channels
django-oauth-toolkit
